Added composite types to the configuration system.
To do this, a new type of options was added: config_struct,
and behavior was added to all functions for handling options
of this generalized type.  The config_struct type is
specialized to a specific composite type by using
the type_definition structure. Registration takes place
in the global array of UserDefinedConfigureTypes.
Added support for postgresql.conf and SET/SHOW commands.
Composite types in the user interface support the syntax:
- option_name = {field_name: field_value, ..., field_name: field_value}
- option_name->array_name[<index>] = value
- option_name = [value, value, ..., value]
- option_name = [<index>: value, ..., <index>: value]
Other features are described in src/backend/utils/misc/README.